Don't blow up when Java 8 JRE is installed.
Previously (#54) I handled the case where Java 8 was installed but this
just looked at the version for the JDK and tried to install the JRE
assuming they had the same version. Instead let's add a
`java_jre_version` fact and query that for the JRE and allow installing
one or the other depending on the installed version.